python-emotion-api/requirements.txt
#필요한 라이브러리 명시
fastapi
uvicorn
pydantic
transformers
torch
python-dotenv
